requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.24.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available
pillow>=10.0.0
numpy>=1.24.0
imagehash>=4.3.1  # For perceptual hashing and duplicate detection

# Database (sqlite3 is built-in to Python)

# Optional quality boosters
sentence-transformers>=2.2.0  # For text encoding if not using CLIP text encoder

# CLI and API
click>=8.1.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
flask>=3.0.0  # For simple search UI

# Utilities
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
filelock>=3.13.0

# Embedding API support
google-genai>=0.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
httpx>=0.24.1  # For FastAPI testing
